+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: moel-asr <moel-asr@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/19 20:06:10 by moel-asr          #+#    #+#             */
-/*   Updated: 2022/10/24 03:18:09 by moel-asr         ###   ########.fr       */
+/*   Updated: 2022/10/25 23:02:03 by moel-asr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 # include <unistd.h>
 
 int	ft_put_unsigned_int(unsigned int n);
-int	ft_putaddress(size_t n);
+int	ft_putaddress(unsigned long n);
 int	ft_putchar(char c);
 int	ft_puthex(unsigned int n, char *arr);
 int	ft_putnbr(int n);
